+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Version 2.0.0](https://github.com/yannickl/YLTCPBroadcaster/releases/tag/2.0.0)
 *Released on 2015-10-29.*
 
+- [ADD] `YLTCPBroadcasterDelegate` property
 - [ADD] `YLTCPSocketStatus` enum
 - [ADD] `maxConcurrentConnectionCount` property
 - [UPDATE] Improve compatibility with Swift 2