+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 ## [Version 2.0.0](https://github.com/yannickl/YLTCPBroadcaster/releases/tag/2.0.0)
-*Released on 2015-10-29.*
+*Released on 2015-11-04.*
 
 - [ADD] `YLTCPBroadcasterDelegate` property
 - [ADD] `YLTCPSocketStatus` enum